refactor: extract navbar partials
## Navbar Partials

Creates a partials folder and a navbar folder inside.

Pulls out the light/dark mode toggle and the github link into their own
partials. It isn't super necessary, but the svgs are big chunks and
clutter up the overall flow of the code. Makes it easier to read and
reason about it all.