fix: honor chosen theme regardless of system preference
## Regard Chosen Theme

If you chose light mode for this site, but your system preference is for
dark mode, the theme specifically chosen for this site should be
honored.

Before, the system preference would override the previously chosen theme
on initial load.

## Add Light Mode Colors To Styles

For the fix, we need specific light mode colors. Extracted the colors
into a mixin and put them in as the default to keep one source of truth.