fix: quirk with light/dark mode
## CSS Swap

There were some CSS values that needed to be swapped to properly enable
light and dark mode when the saved theme doesn't match the system theme.

## Theme Controller

Separately saving the theme to localStorage. The theme choice should
only be saved to local storage if the user clicks the theme toggle. It
should be empty if it's going off the system choice.

We were saving to local storage on first load. Now it only saves the
theme choice if the user clicks the button to make an active choice.
Otherwise local storage has no theme choice.