feat: extract frameworks from language in data file
## Extract Frameworks

Frameworks are sub lists for languages. This pulls them out of the name
string and gives them their own entries. It also adds tags to the data
for the upcoming filter functionality. Now frameworks can have their own
individual tag and there is more flexibility in how we render them.