fix: improper tag filtering match
## Improper Tag Filtering Matches

The tag list was a string when pulled in from the data attribute.
Because of this, the includes was matching any substring in the tag list
string. This caused every job with rails to match ai.

Converting the tag list to an actual JavaScript array fixed the issue.
Now includes checks the array for an actual match, not a substring.